refactor: extract global-state task types
